+++ packages/deployment/upgrade-test/upgrade-test-scripts/agoric-upgrade-10/post.test.js
@@ -1,8 +1,13 @@
 import test from 'ava';
 
-import { agd, agoric } from '../cliHelper.js';
-import { GOV1ADDR, GOV2ADDR, GOV3ADDR, USER1ADDR } from '../constants.js';
-import { calculateWalletState } from '../commonUpgradeHelpers.js';
+import { agd, agoric } from '../lib/cliHelper.js/index.js';
+import {
+  GOV1ADDR,
+  GOV2ADDR,
+  GOV3ADDR,
+  USER1ADDR,
+} from '../lib/constants.js/index.js';
+import { calculateWalletState } from '../lib/commonUpgradeHelpers.js/index.js';
 
 test('DeliverInbound from un-provisioned account is discarded', async t => {
   const result = await agd.query('swingset', 'mailbox', USER1ADDR);
