+++ packages/deployment/upgrade-test/upgrade-test-scripts/lib/cliHelper.js
@@ -1,6 +1,6 @@
 /* eslint-disable @jessie.js/safe-await-separator */
 import { $, execaCommand } from 'execa';
-import { BINARY } from './constants.js';
+import { BINARY } from './constants.js/index.js';
 
 export const executeCommand = async (command, params, options = {}) => {
   const { stdout } = await execaCommand(
